add two more util functions to load the descriptors
Adds from_region to allow a user to build the boot descriptors from a provided buffer instead of utilizing memory-mapped reads at an address. This is useful in case application code has already configured the flexspi module as not all configs will read bits in the same was as the FCB. Because of the pointer arithmetic in locating the app descriptors, this function is considered unsafe as special care MUST be given to the buffer argument to avoid invalid memory access.
Adds from_header_and_app_regions to allow a user to build the boot descriptors from two small buffers containing the bytes for the header and the bytes for the app descriptors separately. This is useful in case these chunks are not contiguous in memory.